Moving registry into sql module
