+++ ibis/backends/base/sql/registry/aggregate.py
