+++ ibis/backends/base/sql/registry/binary_infix.py
