+++ ibis/backends/base/sql/registry/case.py
