+++ ibis/backends/base/sql/registry/literal.py
