+++ ibis/backends/base/sql/registry/string.py
