+++ ibis/backends/base/sql/registry/window.py
