+++ ibis/backends/impala/ddl.py
@@ -14,7 +14,7 @@
 
 import json
 
-from ibis.backends.base import type_to_sql_string
+from ibis.backends.base.sql import type_to_sql_string
 from ibis.backends.base_sql import ddl as base_sql_ddl
 from ibis.backends.base_sql.ddl import (
     AlterTable,
