+++ ibis/backends/impala/udf.py
@@ -20,7 +20,7 @@ import ibis.expr.rules as rlz
 import ibis.expr.signature as sig
 import ibis.udf.validate as v
 import ibis.util as util
-from ibis.backends.base import fixed_arity, sql_type_names
+from ibis.backends.base.sql import fixed_arity, sql_type_names
 
 from . import compiler as comp
